pyproject.toml
[tool.poetry]
name = "vedic"
version = "0.0.1"
description = "An API for vedic text querying."
authors = ["externref <shiva02939@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.100.0"
asyncpg = "^0.28.0"
attrs = "^23.1.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
asyncpg-stubs = "^0.28.0"
isort = "^5.12.0"
black = "^23.7.0"
ruff = "^0.0.278"
pyright = "^1.1.317"

[tool.pyright]
pythonVersion = "3.8"
typeCheckingMode = "strict"
reportPrivateUsage = false
reportImportCycles = false
reportIncompatibleMethodOverride = false

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
line_length = 120
skip = "env"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
